requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
google-cloud-storage==2.18.2
python-dateutil==2.9.0.post0
pydantic==2.9.2
requests==2.32.3
